requirements.txt
click
biopython
requests
openai
